Move PS signatures from crate to zmix
