+++ Cargo.toml
@@ -2,5 +2,4 @@
 members = [
   "libursa",
   "libzmix",
-  "libzmix/ps",
 ]
