--- libzmix/ps/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "ps"
-version = "0.1.0"
-authors = ["lovesh <lovesh.bond@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-rand = "0.7"
-log = "*"
-failure = "0.1.5"
-serde = { version = "1.0", features = ["derive"] }
-ursa = { path = "../../libursa" }
-zmix = { path = ".." }
-
-[dependencies.amcl_wrapper]
-version = "0.1.5"
-default-features = false
-features = ["bls381"]
-
-[features]
-default = ["G1G2"]
-G1G2 = []
-G2G1 = []
